Time: 47 ms (65.28%), Space: 11 MB (30.13%) - LeetHub

+++ 1209-remove-all-adjacent-duplicates-in-string-ii/1209-remove-all-adjacent-duplicates-in-string-ii.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     string removeDuplicates(string s, int k) {
-        vector<pair<char, short>> v;
+        vector<pair<char, int>> v;
             for(auto it:s){
                     if(v.empty() || v.back().first!=it){
                             v.push_back({it,0});
